Refactor: Remove unused code for NativeRuntimeBuilder, OutputPathSafety, and related tests; update .editorconfig for enhanced naming rules and formatting consistency
