+++ src/InfiniFrame.Tools.Pack/CommandLine.cs
@@ -1,7 +1,9 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
-namespace InfiniFrame.Tools.Pack.Services;
+using InfiniFrame.Tools.Pack.Services;
+
+namespace InfiniFrame.Tools.Pack;
 // -----------------------------------------------------------------------------------------------------------------
 // Methods
 // -----------------------------------------------------------------------------------------------------------------
@@ -45,7 +47,7 @@ internal static class CommandLine {
         Logger.Information("InfiniFrame.Pack");
         Logger.Information("Usage:");
         Logger.Information("  infiniframe-pack publish <project.csproj> [options]");
-        Logger.Information(string.Empty);
+        Logger.Information("");
         Logger.Information("Options:");
         Logger.Information("  --rid <RID|auto>             Runtime identifier. Default: auto");
         Logger.Information("  --configuration <Config>      Build configuration. Default: Release");
@@ -59,7 +61,7 @@ internal static class CommandLine {
         Logger.Information("                                Explicit fallback directory for native artifacts");
         Logger.Information("  --allow-stale-native-fallback");
         Logger.Information("                                Allow using fallback artifacts when preflight fails validation");
-        Logger.Information(string.Empty);
+        Logger.Information("");
         Logger.Information("Environment overrides:");
         Logger.Information("  {FallbackEnvVar}=<path>", NativeArtifactsFallbackPathEnvVar);
         Logger.Information("  {AllowStaleEnvVar}=true|false", AllowStaleNativeFallbackEnvVar);
