+++ src/InfiniFrame.Tools.Pack/Resolvers/MsBuildPropertyResolver.cs
@@ -3,7 +3,7 @@
 // ---------------------------------------------------------------------------------------------------------------------
 using System.Diagnostics;
 
-namespace InfiniFrame.Tools.Pack.Services;
+namespace InfiniFrame.Tools.Pack.Resolvers;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
