+++ src/InfiniFrame.Tools.Pack/Resolvers/ProjectInfoResolver.cs
@@ -1,7 +1,7 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
-namespace InfiniFrame.Tools.Pack.Services;
+namespace InfiniFrame.Tools.Pack.Resolvers;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
