+++ src/InfiniFrame.Tools.Pack/Resolvers/RuntimeResolver.cs
@@ -3,7 +3,7 @@
 // ---------------------------------------------------------------------------------------------------------------------
 using System.Runtime.InteropServices;
 
-namespace InfiniFrame.Tools.Pack.Services;
+namespace InfiniFrame.Tools.Pack.Resolvers;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
