+++ tests/InfiniFrameTests.Tools.Pack/CommandLineTests.cs
@@ -1,9 +1,10 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
+using InfiniFrame.Tools.Pack;
 using InfiniFrame.Tools.Pack.Services;
 
-namespace InfiniFrameTests.Tools.Pack.Services;
+namespace InfiniFrameTests.Tools.Pack;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
