+++ tests/InfiniFrameTests.Tools.Pack/Resolvers/ProjectInfoResolverTests.cs
@@ -1,10 +1,10 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
-using InfiniFrame.Tools.Pack.Services;
+using InfiniFrame.Tools.Pack.Resolvers;
 using InfiniFrameTests.Tools.Pack.TestUtilities;
 
-namespace InfiniFrameTests.Tools.Pack.Services;
+namespace InfiniFrameTests.Tools.Pack.Resolvers;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
