+++ tests/InfiniFrameTests.Tools.Pack/Resolvers/RuntimeResolverTests.cs
@@ -1,10 +1,10 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
-using InfiniFrame.Tools.Pack.Services;
+using InfiniFrame.Tools.Pack.Resolvers;
 using System.Runtime.InteropServices;
 
-namespace InfiniFrameTests.Tools.Pack.Services;
+namespace InfiniFrameTests.Tools.Pack.Resolvers;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
