+++ tests/InfiniFrameTests.Tools.Pack/Services/ParseResultTests.cs
@@ -1,6 +1,7 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
+using InfiniFrame.Tools.Pack;
 using InfiniFrame.Tools.Pack.Services;
 
 namespace InfiniFrameTests.Tools.Pack.Services;
