+++ tests/InfiniFrameTests.Tools.Pack/Services/PublishOutputCleanerTests.cs
@@ -36,7 +36,7 @@ public class PublishOutputCleanerTests {
 
         Directory.CreateDirectory(wwwroot);
         await File.WriteAllTextAsync(Path.Join(wwwroot, "index.html"), "<html></html>");
-        foreach (string file in NativeRuntimeBuilder.NativeRuntimeFiles) {
+        foreach (string file in PublishOutputCleaner.NativeRuntimeFiles) {
             await File.WriteAllTextAsync(Path.Join(output, file), string.Empty);
         }
 
@@ -46,7 +46,7 @@ public class PublishOutputCleanerTests {
         // Assert
         await Assert.That(warnings.Length).IsEqualTo(0);
         await Assert.That(Directory.Exists(wwwroot)).IsFalse();
-        foreach (string file in NativeRuntimeBuilder.NativeRuntimeFiles) {
+        foreach (string file in PublishOutputCleaner.NativeRuntimeFiles) {
             await Assert.That(File.Exists(Path.Join(output, file))).IsFalse();
         }
     }
@@ -68,7 +68,7 @@ public class PublishOutputCleanerTests {
     public async Task Cleanup_ReturnsWarning_WhenNativeArtifactDeletionFails() {
         // Arrange
         string output = TemporaryDirectory.Path;
-        string nativeArtifactPath = Path.Join(output, NativeRuntimeBuilder.NativeRuntimeFiles[0]);
+        string nativeArtifactPath = Path.Join(output, PublishOutputCleaner.NativeRuntimeFiles[0]);
         await File.WriteAllTextAsync(nativeArtifactPath, "locked");
         File.SetAttributes(nativeArtifactPath, File.GetAttributes(nativeArtifactPath) | FileAttributes.ReadOnly);
